+++ src/routes/Layout.svelte
@@ -11,12 +11,13 @@
   import Sidebar from "./Sidebar.svelte";
   let leaveModal = false;
   const dispatch = createEventDispatcher();
+  export let view;
   export let profileQuestionsAnswered = 0;
   export let totalProfileQuestions = 7;
 </script>
 
 <Layout>
-  <Sidebar {totalProfileQuestions} {profileQuestionsAnswered} on:change-view on:leave-rally={() => {
+  <Sidebar {view} {totalProfileQuestions} {profileQuestionsAnswered} on:change-view on:leave-rally={() => {
     leaveModal = true;
   }} />
   <Body>