+++ src/routes/Main.svelte
@@ -14,15 +14,19 @@ import demographicsSchema from './demographics/survey-schema';
 import { questionIsAnswered } from './demographics/survey-tools';
 
 const store = getContext("rally:store");
+const notification = getContext("rally:notification");
 
 let view = 'current-studies';
 function changeView(event) {
-    view = event.detail;
+    view = event.detail.view;
+    if (!event.detail.suppressNotifications) {
+        notification.clear();
+    }
     window.scrollTo(0, 0);
 }
 
-function joinStudy(studyID) { store.updateStudyEnrollment(studyID, true); }
-function leaveStudy(studyID) { store.updateStudyEnrollment(studyID, false); }
+function joinStudy(studyID) { store.updateStudyEnrollment(studyID, true); notification.send({code: "SUCCESSFULLY_JOINED_STUDY"}); }
+function leaveStudy(studyID) { store.updateStudyEnrollment(studyID, false); notification.send({code: "SUCCESSFULLY_LEFT_STUDY"}); }
 
 /* ------------------------------ PROFILE COMPLETION ------------------------------ */
 // get the number of profile questions answered
@@ -38,15 +42,16 @@ onMount(() => { mounted = true; })
 
 {#if mounted}
 
-<Layout 
+<Layout
     on:change-view={changeView}
     on:leave-rally
+    {view}
     {profileQuestionsAnswered}
     {totalProfileQuestions}
 >
         {#if view === 'manage-profile'}
             <MainContent>
-                <Demographics />
+                <Demographics on:redirect-to={changeView} />
             </MainContent>
         {:else if view === 'privacy-notice'}
             <MainContent>
@@ -58,6 +63,9 @@ onMount(() => { mounted = true; })
                     <CurrentStudies
                         sidebarOffset
                         studies={$store.availableStudies}
+                        on:cta-clicked={() => {
+                            notification.clear();
+                        }}
                         on:join-study={(evt) => { joinStudy(evt.detail); }}
                         on:leave-study={(evt) => { leaveStudy(evt.detail); }} />
                 </MainContent>