+++ src/routes/current-studies/StudyCard.svelte
@@ -1,42 +1,3 @@
-<script context=module>
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
- // ^^^ the linter currently forces this block even though it appears below.
- // see https://github.com/mozilla-rally/core-addon/issues/184.
-import { writable, get } from 'svelte/store';
-import irb from "../irbs/";
-import IRBWindow from "../irbs/IRBWindow.svelte";
-
-let notificationID = writable(undefined);
-let whichNotification = writable(false);
-let activeKey = writable(undefined);
-
-let firstRuns = {};
-
- function showNotification(joinOrLeave, key) {
-    // absorb first run calls of showNotification.
-    // This enables future reactive updates to trigger.
-    if (!(key in firstRuns)) {
-        firstRuns[key] = true;
-    } else {
-        const nid = get(notificationID);
-        activeKey.set(key);
-        whichNotification.set(joinOrLeave);
-        if (nid) {
-            clearTimeout(get(notificationID));
-            notificationID.set(undefined);
-        }
-        notificationID.set(setTimeout(() => {
-            activeKey.set(undefined);
-            whichNotification.set(undefined);
-            notificationID.set(undefined);
-        }, 3000));
-    }
-}
-
-</script>
-
 <script> 
  /* This Source Code Form is subject to the terms of the Mozilla Public
   * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -55,8 +16,8 @@ import StudyCard from '../../components/study-card/StudyCard.svelte';
 import StudyCardHeader from '../../components/study-card/Header.svelte';
 import Button from '../../components/Button.svelte';
 import Dialog from '../../components/Dialog.svelte';
-import SuccessfullyJoinedStudyNotification from './SuccessfullyJoinedStudyNotification.svelte';
-import SuccessfullyLeftStudyNotification from './SuccessfullyLeftStudyNotification.svelte';
+import IRBWindow from '../irbs/IRBWindow.svelte';
+import irb from "../irbs";
 
 export let joined = false;
 export let imageSrc;
@@ -70,13 +31,8 @@ export let privacyPolicyLink;
 export let tags;
 export let detailsDirectName;
 export let detailsDirectLink;
-export let sidebarOffset = false; // sidebar offset for notifications
 
 const dispatch = createEventDispatcher();
-
-const key =
-    `modal-${Math.random().toString(36).substring(2, 15) +
-    Math.random().toString(36).substring(2, 15)}`;
 
 let joinModal = false;
 
@@ -85,10 +41,6 @@ function triggerJoinEvent() {
     dispatch('cta-clicked');
     joinModal = true;
 }
-
-// when the joined part changes, do something.
-$: showNotification(joined, key);
-$: isActive = $activeKey !== undefined && $activeKey === key;
 
 </script>
 
@@ -189,30 +141,4 @@ $: isActive = $activeKey !== undefined && $activeKey === key;
         </Button>
     </div>
     </Dialog>
-{/if}
-
-<!-- 
-    this simple notification system should suffice for GTM.
-    The component context controls whether or not any study cards
-    show a notification. If a study card status changes (joined / unjoined)
-    then all other notifications are removed, leaving room for the newly-changed
-    study card to have its notification visible.
-
-    NOTE: There is an edge case where I cannot get this notification to hide when attempting
-    to suppress it when joinModal = true. From a UX point of view it should be fine enough,
-    but leaving this note here to investigate if this is a Svelte edge case or otherwise.
- -->
-{#if isActive && $notificationID !== undefined}
-    {#key $notificationID}
-        {#if $whichNotification === true}
-        <SuccessfullyJoinedStudyNotification 
-            location={sidebarOffset ? "top-left" : "top"}
-            xOffset={sidebarOffset ? "var(--main-notification-offset)" : undefined } />
-        {:else}
-        <SuccessfullyLeftStudyNotification
-            location={sidebarOffset ? "top-left" : "top"}
-            xOffset={sidebarOffset ? "var(--main-notification-offset)" : undefined }
-            />
-        {/if}
-    {/key}
 {/if}